backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
python-multipart==0.0.19
pymupdf==1.25.1
python-docx==1.1.2
openai==1.58.1
tiktoken==0.8.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1
alembic==1.14.0
psycopg2-binary==2.9.10
